pyproject.toml
[project]
name = "app"
version = "1.0.0"
authors = [
    { name = "Roman", email = "play62342@gmail.com" },
]
# https://docs.astral.sh/uv/reference/resolver-internals/#requires-python
requires-python = "==3.13.*"
# https://docs.astral.sh/uv/concepts/dependencies/#project-dependencies
dependencies = [
    "Jinja2==3.1.*",
    "fastapi==0.115.*",
    "uvicorn==0.34.*",
    "gunicorn==23.0.*",
    "SQLAlchemy==2.0.*",
    "alembic==1.14.*",
    "asyncpg==0.30.*",
    "fastapi-utils>=0.8.0",
    "typing-inspect>=0.9.0",
    "sqladmin==0.20.1",
    "itsdangerous==2.2.0",
    "pydantic-settings>=2.7.0",
    "aiohttp>=3.11.18",
    "pyjwt>=2.10.1",
]

# https://docs.astral.sh/uv/concepts/dependencies/#development-dependencies
[dependency-groups]
dev = [
    "grpcio-tools==1.66.*",
    "pytest==8.3.*",
    "pynvim",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src"]

[tool.hatch.build.targets.wheel.sources]
"src" = "src"
